file deletion and save bug fix

+++ utils/clearTemporaryDirecotry.jsx
@@ -1,11 +1,13 @@
 import * as FileSystem from 'expo-file-system';
 
-const ClearTemporaryDirectory = async (directoryPath) => {
+const ClearTemporaryDirectory = async (directoryPath, fileName) => {
     try {
         const files = await FileSystem.readDirectoryAsync(directoryPath);
         for (const file of files) {
-            const filePath = `${directoryPath}/${file}`;
-            await FileSystem.deleteAsync(filePath);
+            if (file !== fileName) {
+                const filePath = `${directoryPath}/${file}`;
+                await FileSystem.deleteAsync(filePath);
+            }
         }
     } catch (error) {
         console.error('Error clearing directory:', error);
